src/inijira/requirements/base.txt
click
python-dotenv

beautifulsoup4
lxml
